requirements.txt
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.29.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.1
numpy==1.22.2
packaging==21.3
pandas==1.4.0
Pillow==9.0.1
pyparsing==3.0.7
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
Werkzeug==2.0.2
